Extract CharBuffer-based CharSequence to a file
